requirements.txt
certifi==2023.11.17
contourpy==1.2.0
cycler==0.12.1
fonttools==4.49.0
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pillow==10.2.0
polygon-api-client==1.13.4
pre-commit==3.6.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
seaborn==0.13.2
six==1.16.0
tzdata==2024.1
urllib3==1.26.18
websockets==12.0
